requirements.txt
# Core data science
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0

# Web scraping
google-play-scraper>=1.2.0

# NLP and sentiment analysis
transformers>=4.30.0
torch>=2.0.0
vaderSentiment>=3.3.2
textblob>=0.17.1
nltk>=3.8.0
spacy>=3.5.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
wordcloud>=1.9.0

